% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/show_connection.R
\name{show_connection}
\alias{show_connection}
\title{Show connection parameters.}
\usage{
show_connection()
}
\description{
The function shows the current values of the connection parameters
\code{url, user, password, verify} and \code{encryption}.
They are used to arrange a connection with ATSD.
}
\seealso{
You could change the connection parameters with the \code{\link{set_connection}}
function and save that changes to the configuration file with
the \code{\link{save_connection}} function.
}

