% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_zoo.R
\name{to_zoo}
\alias{to_zoo}
\title{Build zoo object from data frame.}
\usage{
to_zoo(dfr, timestamp = "Timestamp", value = "Value")
}
\arguments{
\item{dfr}{The data frame with columns for time stamps and for values.}

\item{timestamp}{Name or number of a column with time stamps. By default, \code{timestamp = "Timestamp"}.}

\item{value}{Vector of names or numbers of columns with series values. 
By default, \code{value = "Value"}.}
}
\description{
The function builds a zoo object from given data frame. 
The \code{timestamp} argument provides
a column of the data frame which is used as index for the zoo object.
The \code{value} argument gives series to be saved in the zoo object.
If several columns are listed in \code{value} argument the multivariate
zoo object will be built.
Information from other columns is ignored.
To use this function the 'zoo' package should be installed.
To install the 'zoo' package type:
\code{install.packages("zoo")}.
}
