% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/to_zoo.R
\name{to_zoo}
\alias{to_zoo}
\title{Build zoo object from data frame.}
\usage{
to_zoo(dfr)
}
\arguments{
\item{dfr}{The data frame with 'Timestamp' and 'Value' columns.}
}
\description{
The function builds a zoo object based on the 'Timestamp' and 'Value'
columns of the given data frame. Information from other columns will be lost.
To use this function the 'zoo' package should be installed.
To install the 'zoo' package type:
\code{install.packages("zoo")}.
}

